Share param form between json-rpc form and contract form
